+++ packages/office-ui-fabric-react/src/components/DetailsList/DetailsList.test.tsx
@@ -312,6 +312,11 @@ describe('DetailsList', () => {
       );
     };
 
+    const getCellValueKey = (item: any, index: number, column: IColumn) => {
+      const valueKey = item && column && column.fieldName ? item[column.fieldName] : column.key + index;
+      return valueKey;
+    };
+
     jest.useFakeTimers();
 
     let component: any;
@@ -324,6 +329,7 @@ describe('DetailsList', () => {
         // tslint:disable-next-line:jsx-no-lambda
         onShouldVirtualize={() => false}
         onRenderItemColumn={onRenderColumn}
+        getCellValueKey={getCellValueKey}
       />
     );
 
@@ -391,25 +397,18 @@ describe('DetailsList', () => {
   });
 
   it('invokes optional onColumnResize callback per IColumn if defined when columns are adjusted', () => {
-    const detailsList = mount(
-      <DetailsList
-        items={mockData(2)}
-        skipViewportMeasures={true}
-        // tslint:disable-next-line:jsx-no-lambda
-        onShouldVirtualize={() => false}
-      />
-    );
-
     const columns: IColumn[] = mockData(2, true);
     columns[0].onColumnResize = jest.fn();
     columns[1].onColumnResize = jest.fn();
 
-    // componentWillReceiveProps not executed on initial render in test
-    // so we need to force one via setProps and update.
-    const newProps = { columns };
-
-    detailsList.setProps(newProps);
-    detailsList.update();
+    mount(
+      <DetailsList
+        items={mockData(2)}
+        columns={columns}
+        // tslint:disable-next-line:jsx-no-lambda
+        onShouldVirtualize={() => false}
+      />
+    );
 
     expect(columns[0].onColumnResize).toHaveBeenCalledTimes(1);
     expect(columns[1].onColumnResize).toHaveBeenCalledTimes(1);