+++ packages/office-ui-fabric-react/src/components/DetailsList/DetailsRow.base.tsx
@@ -166,8 +166,9 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowBaseProps, IDetails
       onRenderCheck = this._onRenderCheck,
       onRenderDetailsCheckbox,
       onRenderItemColumn,
+      getCellValueKey,
       selectionMode,
-      viewport,
+      rowWidth = 0,
       checkboxVisibility,
       getRowAriaLabel,
       getRowAriaDescribedBy,
@@ -178,6 +179,7 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowBaseProps, IDetails
       selection,
       indentWidth,
       shimmer,
+      enableUpdateAnimations,
       compact,
       theme,
       styles,
@@ -185,7 +187,7 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowBaseProps, IDetails
     } = this.props;
     const { columnMeasureInfo, isDropping, groupNestingDepth } = this.state;
     const { isSelected = false, isSelectionModal = false } = this.state.selectionState as IDetailsRowSelectionState;
-    const isDraggable = Boolean(dragDropEvents && dragDropEvents.canDrag && dragDropEvents.canDrag(item));
+    const isDraggable = dragDropEvents ? !!(dragDropEvents.canDrag && dragDropEvents.canDrag(item)) : undefined;
     const droppingClassName = isDropping ? (this._droppingClassNames ? this._droppingClassNames : DEFAULT_DROPPING_CSS_CLASS) : '';
     const ariaLabel = getRowAriaLabel ? getRowAriaLabel(item) : undefined;
     const ariaDescribedBy = getRowAriaDescribedBy ? getRowAriaDescribedBy(item) : undefined;
@@ -204,7 +206,8 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowBaseProps, IDetails
         checkboxCellClassName,
         droppingClassName,
         className,
-        compact
+        compact,
+        enableUpdateAnimations
       })
     };
 
@@ -214,6 +217,7 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowBaseProps, IDetails
       shimmerIconPlaceholder: this._classNames.shimmerIconPlaceholder,
       shimmer: this._classNames.shimmer,
       cell: this._classNames.cell,
+      cellAnimation: this._classNames.cellAnimation,
       cellPadded: this._classNames.cellPadded,
       cellUnpadded: this._classNames.cellUnpadded,
       fields: this._classNames.fields
@@ -237,6 +241,7 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowBaseProps, IDetails
         columnStartIndex={showCheckbox ? 1 : 0}
         onRenderItemColumn={onRenderItemColumn}
         shimmer={shimmer}
+        getCellValueKey={getCellValueKey}
       />
     );
 
@@ -252,6 +257,12 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowBaseProps, IDetails
     return (
       <FocusZone
         {...getNativeProps(this.props, divProperties)}
+        {...(typeof isDraggable === 'boolean'
+          ? {
+              'data-is-draggable': isDraggable, // This data attribute is used by some host applications.
+              draggable: isDraggable
+            }
+          : {})}
         direction={FocusZoneDirection.horizontal}
         ref={this._onRootRef}
         componentRef={this._focusZone}
@@ -263,10 +274,8 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowBaseProps, IDetails
         data-selection-index={itemIndex}
         data-item-index={itemIndex}
         aria-rowindex={itemIndex + 1}
-        data-is-draggable={isDraggable}
-        draggable={isDraggable}
         data-automationid="DetailsRow"
-        style={{ minWidth: viewport ? viewport.width : 0 }}
+        style={{ minWidth: rowWidth }}
         aria-selected={ariaSelected}
         allowFocusRoot={true}
       >
@@ -301,6 +310,7 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowBaseProps, IDetails
               itemIndex={itemIndex}
               columnStartIndex={(showCheckbox ? 1 : 0) + columns.length}
               onRenderItemColumn={onRenderItemColumn}
+              getCellValueKey={getCellValueKey}
             />
           </span>
         )}