+++ packages/office-ui-fabric-react/src/components/Panel/Panel.base.tsx
@@ -121,7 +121,7 @@ export class PanelBase extends BaseComponent<IPanelProps, IPanelState> implement
     const isOnRightSide = isRTL ? isLeft : !isLeft;
     const headerTextId = headerText && id + '-headerText';
     const customWidthStyles = type === PanelType.custom || type === PanelType.customNear ? { width: customWidth } : {};
-    const nativeProps = getNativeProps(this.props, divProperties);
+    const nativeProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(this.props, divProperties);
 
     const isOpen = this._isOpen;
 
@@ -136,8 +136,8 @@ export class PanelBase extends BaseComponent<IPanelProps, IPanelState> implement
       hasCloseButton,
       headerClassName,
       isAnimating,
-      isFooterAtBottom,
       isFooterSticky,
+      isFooterAtBottom,
       isOnRightSide,
       isOpen,
       isHiddenOnDismiss,