+++ packages/office-ui-fabric-react/src/components/TextField/TextField.test.tsx
@@ -57,7 +57,13 @@ describe('TextField', () => {
     expect(tree).toMatchSnapshot();
   });
 
-  it('renders multiline TextField correctly with props affecting styling', () => {
+  it('renders multiline with placeholder correctly', () => {
+    const component = renderer.create(<TextField label="Label" multiline={true} placeholder="test placeholder" />);
+    const tree = component.toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+
+  it('renders multiline correctly with props affecting styling', () => {
     const component = renderer.create(
       <TextField label="Label" errorMessage={'test message'} underlined={true} prefix={'test prefix'} suffix={'test suffix'} />
     );