+++ apps/fabric-website-resources/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @uifabric/fabric-website-resources
 
-This log was last generated on Wed, 12 Jun 2019 22:15:55 GMT and should not be manually modified.
+This log was last generated on Mon, 22 Jul 2019 12:21:49 GMT and should not be manually modified.
+
+## 6.19.1
+Mon, 22 Jul 2019 12:21:49 GMT
+
+### Patches
+
+- Adding @types/react and @types/react-dom to package.json that have peer dependencies on react and react-dom (backport of #9798).
+
+## 6.19.0
+Wed, 10 Jul 2019 00:26:15 GMT
+
+### Minor changes
+
+- Add DetailsList animation example page
 
 ## 6.18.5
 Wed, 12 Jun 2019 22:15:55 GMT