+++ packages/date-time/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @uifabric/date-time
 
-This log was last generated on Wed, 12 Jun 2019 22:15:55 GMT and should not be manually modified.
+This log was last generated on Mon, 22 Jul 2019 12:21:49 GMT and should not be manually modified.
+
+## 6.3.6
+Mon, 22 Jul 2019 12:21:49 GMT
+
+### Patches
+
+- Adding @types/react and @types/react-dom to package.json that have peer dependencies on react and react-dom (backport of #9798).
+
+## 6.3.5
+Tue, 16 Jul 2019 12:23:13 GMT
+
+### Patches
+
+- update snapshots
 
 ## 6.3.4
 Wed, 12 Jun 2019 22:15:55 GMT